[schedule] Fix data races in delayed call connections.
Given two threads scheduling calls and disconnecting their connections:

T0                            | T1
connection = delayed_call(…); | connection = delayed_call(…);
connection.disconnect();      | connection.disconnect()

Both delayed calls will create a connection to the same signal, for
which the locks are okay. During the setup of the delayed call a check
was done to test if the signal was empty, which meant testing all
connections. But then, the call to disconnect() would modify one
connection to the signal. If one thread disconnects while the other is
checking if the signal is empty, we have a data race.

This commit introduce a connection type specific to the
delayed_call_manager, which redirects connection operations to it in
order to be guarded by its mutex.